Distributed contents of util/dsutil.py to api/ and util/
