+++ xcube/api/__init__.py
@@ -2,6 +2,6 @@ from .api import XCubeAPI
 from .chunk import chunk_dataset
 from .dump import dump_dataset
 from .new import new_cube
-from .point import get_cube_point_indexes, get_cube_values_for_indexes, get_cube_values_for_points, get_dataset_indexes
+from .extract import get_cube_point_indexes, get_cube_values_for_indexes, get_cube_values_for_points, get_dataset_indexes
 from .readwrite import open_dataset, read_dataset, write_dataset
 from .verify import assert_cube, verify_cube
