+++ xcube/api/gen/default/iproc.py
@@ -23,8 +23,8 @@ from typing import Tuple
 
 import xarray as xr
 
-from xcube.util.dsutil import get_time_in_days_since_1970
 from xcube.util.constants import CRS_WKT_EPSG_4326
+from xcube.util.timecoord import get_time_in_days_since_1970
 from ..iproc import InputProcessor, register_input_processor, ReprojectionInfo
