+++ xcube/api/select.py
@@ -1,0 +1,19 @@
+from typing import Sequence
+
+import xarray as xr
+
+
+def select_vars(dataset: xr.Dataset, var_names: Sequence[str] = None) -> xr.Dataset:
+    """
+    Select data variable from given *dataset* and create new dataset.
+
+    :param dataset: The dataset from which to select variables.
+    :param var_names: The names of data variables to select.
+    :return: A new dataset. It is empty, if *var_names* is empty. It is *dataset*, if *var_names* is None.
+    """
+    if var_names is None:
+        return dataset
+    dropped_variables = set(dataset.data_vars.keys()).difference(var_names)
+    if not dropped_variables:
+        return dataset
+    return dataset.drop(dropped_variables)
